Move the analysis-sample fixture one level up
